hf-worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.0
transformers==4.44.0
Pillow==10.4.0
supabase==2.7.2
sentencepiece==0.2.0
protobuf==5.27.3
